fix: refactor client package structure
- Renamed `client` package to `clients` for consistency.
- Updated imports across multiple classes to reflect the new package name.
- Introduced a new service for handling client categories.
- Changed request model to use a string for client category instead of an object.
- Removed unnecessary validation annotations from the Client entity.
