+++ src/main/java/fr/iut/pathpilotapi/bootstrap/BootstrapService.java
@@ -5,8 +5,8 @@
 
 package fr.iut.pathpilotapi.bootstrap;
 
-import fr.iut.pathpilotapi.client.ClientCategory;
-import fr.iut.pathpilotapi.client.repository.ClientCategoryRepository;
+import fr.iut.pathpilotapi.clients.ClientCategory;
+import fr.iut.pathpilotapi.clients.repository.ClientCategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
