+++ src/main/java/fr/iut/pathpilotapi/clients/repository/ClientRepository.java
@@ -3,9 +3,9 @@
  * IUT de Rodez, no author rights
  */
 
-package fr.iut.pathpilotapi.client.repository;
+package fr.iut.pathpilotapi.clients.repository;
 
-import fr.iut.pathpilotapi.client.Client;
+import fr.iut.pathpilotapi.clients.Client;
 import fr.iut.pathpilotapi.salesman.Salesman;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
