+++ src/main/java/fr/iut/pathpilotapi/itineraries/ItineraryRestController.java
@@ -9,7 +9,7 @@ import fr.iut.pathpilotapi.itineraries.dto.ItineraryRequestModel;
 import fr.iut.pathpilotapi.itineraries.dto.ItineraryResponseModel;
 import fr.iut.pathpilotapi.itineraries.dto.ItineraryPagedModelAssembler;
 import fr.iut.pathpilotapi.itineraries.dto.ItineraryResponseModelAssembler;
-import fr.iut.pathpilotapi.itineraries.routes.Route;
+import fr.iut.pathpilotapi.routes.Route;
 import fr.iut.pathpilotapi.salesman.Salesman;
 import fr.iut.pathpilotapi.security.SecurityUtils;
 import io.swagger.v3.oas.annotations.Operation;
