+++ src/test/java/fr/iut/pathpilotapi/clients/ClientCategoryRepositoryIntegrationTest.java
@@ -1,6 +1,6 @@
-package fr.iut.pathpilotapi.client;
+package fr.iut.pathpilotapi.clients;
 
-import fr.iut.pathpilotapi.client.repository.ClientCategoryRepository;
+import fr.iut.pathpilotapi.clients.repository.ClientCategoryRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
