+++ src/test/java/fr/iut/pathpilotapi/clients/ClientServiceIntegrationTest.java
@@ -1,6 +1,6 @@
-package fr.iut.pathpilotapi.client;
+package fr.iut.pathpilotapi.clients;
 
-import fr.iut.pathpilotapi.client.repository.ClientRepository;
+import fr.iut.pathpilotapi.clients.repository.ClientRepository;
 import fr.iut.pathpilotapi.salesman.Salesman;
 import fr.iut.pathpilotapi.salesman.SalesmanRepository;
 import fr.iut.pathpilotapi.test.IntegrationTestUtils;
