+++ src/test/java/fr/iut/pathpilotapi/itineraries/routes/RouteTest.java
@@ -1,6 +1,7 @@
 package fr.iut.pathpilotapi.itineraries.routes;
 
 import fr.iut.pathpilotapi.itineraries.dto.ClientDTO;
+import fr.iut.pathpilotapi.routes.Route;
 import org.junit.jupiter.api.Test;
 import org.springframework.data.mongodb.core.geo.GeoJsonPoint;
