+++ src/test/java/fr/iut/pathpilotapi/services/ClientServiceTest.java
@@ -5,11 +5,11 @@
 
 package fr.iut.pathpilotapi.services;
 
-import fr.iut.pathpilotapi.client.Client;
-import fr.iut.pathpilotapi.client.ClientCategory;
-import fr.iut.pathpilotapi.client.ClientService;
-import fr.iut.pathpilotapi.client.repository.ClientCategoryRepository;
-import fr.iut.pathpilotapi.client.repository.ClientRepository;
+import fr.iut.pathpilotapi.clients.Client;
+import fr.iut.pathpilotapi.clients.ClientCategory;
+import fr.iut.pathpilotapi.clients.ClientService;
+import fr.iut.pathpilotapi.clients.repository.ClientCategoryRepository;
+import fr.iut.pathpilotapi.clients.repository.ClientRepository;
 import fr.iut.pathpilotapi.salesman.Salesman;
 import fr.iut.pathpilotapi.test.IntegrationTestUtils;
 import org.junit.jupiter.api.BeforeEach;
